+++ adapter/crates/lldb/src/sb/sbfile.rs
@@ -4,19 +4,23 @@ cpp_class!(pub unsafe struct SBFile as "SBFile");
 
 unsafe impl Send for SBFile {}
 
-#[repr(C)]
-pub struct FILE;
-
 impl SBFile {
     pub fn new() -> SBFile {
         cpp!(unsafe [] -> SBFile as "SBFile" { return SBFile(); })
     }
-    pub fn from(fd: impl std::os::unix::io::IntoRawFd, mode: &str) -> SBFile {
-        let fd = fd.into_raw_fd();
-        with_cstr(mode, |mode| {
-            cpp!(unsafe [fd as "int", mode as "const char*"] -> SBFile as "SBFile" {
-                return SBFile(fd, mode, true);
-            })
+    #[cfg(unix)]
+    pub fn from(obj: impl std::os::unix::io::IntoRawFd, write: bool) -> SBFile {
+        SBFile::from_impl(obj.into_raw_fd(), write)
+    }
+    #[cfg(windows)]
+    pub fn from(obj: impl std::os::windows::io::IntoRawHandle, write: bool) -> SBFile {
+        let flags = if write { 0 } else { libc::O_RDONLY };
+        let fd = unsafe { libc::open_osfhandle(obj.into_raw_handle() as libc::intptr_t, flags) };
+        SBFile::from_impl(fd, write)
+    }
+    fn from_impl(fd: c_int, write: bool) -> SBFile {
+        cpp!(unsafe [fd as "int", write as "bool"] -> SBFile as "SBFile" {
+            return SBFile(fd, write ? "w" : "r", true);
         })
     }
     pub fn close(&self) -> SBError {